pyproject.toml
[project]
name = "cognitionflow"
version = "0.1.0"
description = "Multi-agent RCA pipeline with persistent vector memory (AutoGen, ChromaDB)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy<2.0.0",
    "pyautogen==0.2.28",
    "chromadb",
    "sentence-transformers",
    "matplotlib",
    "seaborn",
    "polars",
    "pandas",
    "python-dotenv>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
select = ["E", "F", "I"]
ignore = ["E501"]
